Remove report snapshot information
This method of reporting has long been deprecated and is no longer in use.

Some of these have been replaced (e.g. vacuum report), whilst others do
not yet have a replacement (e.g. sequence report). Where replacements
are not yet present in the snapshot, we do not anticipate to reuse this
format, hence removing it for now.